requirements.txt
aiohttp==1.2.0
alembic==0.8.8
appnope==0.1.0
cffi==1.8.3
click==6.6
click-datetime==0.2
cryptography==1.7.1
decorator==4.0.10
elasticsearch==5.4.0
filedepot==0.4.0
Flask==0.11.1
Flask-Alembic==2.0.1
Flask-Cors==3.0.2
Flask-Elasticsearch==0.2.5
Flask-GraphQL==1.3.0
Flask-Login==0.4.0
flask-shell-ipython==0.2.2
Flask-SQLAlchemy==2.1
#graphene==1.0.2
#graphene-sqlalchemy==1.0
#graphql-core==1.0
graphql-relay==0.4.4
gunicorn==19.6.0
idna==2.1
ipython==5.1.0
ipython-genutils==0.1.0
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
passlib==1.7.0
pdfminer.six==20170419
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
promise==0.4.2
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
python-docx==0.8.6
python-editor==1.0.1
pytz==2016.10
reportlab==3.3.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.0
SQLAlchemy-Utils==0.32.9
#tabulate==0.7.7
texttable==0.8.7
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.11
z3c.rml==3.1.0
git+https://github.com/hivdb/sierra-client.git#egg=sierrapy&subdirectory=python
git+https://github.com/philiptzou/graphene-sqlalchemy.git@patch-1#egg=graphene-sqlalchemy==1.0.1dev0
git+https://github.com/philiptzou/graphene.git@patch-1#egg=graphene==1.0.3dev0
git+https://github.com/philiptzou/graphql-core.git@patch-1#egg=graphql-core==1.0.1dev0
